# spec/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"AccessTokenResponse": {
				"properties": {
					"access_token": {
						"type": "string"
					},
					"token_type": {
						"type": "string"
					},
					"scope": {
						"type": "string"
					},
					"expires_in": {
						"type": "number",
						"format": "double"
					},
					"refresh_token": {
						"type": "string"
					}
				},
				"required": [
					"access_token",
					"token_type",
					"scope",
					"expires_in"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "gm-companion-backend",
		"version": "0.2.0",
		"description": "A server backend for the gm-companion.",
		"license": {
			"name": "AGPL-3.0-or-later"
		},
		"contact": {
			"name": "Phil Hoffmann"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/spotify/login": {
			"get": {
				"operationId": "Login",
				"responses": {
					"302": {
						"description": "Redirect"
					}
				},
				"description": "Start the login process by redirecting the user to the spotify login page.",
				"security": [],
				"parameters": [
					{
						"description": "URI that the user will be redirected to, after the login procedure is finished.",
						"in": "query",
						"name": "redirect_uri",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/spotify/token": {
			"get": {
				"operationId": "Token",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AccessTokenResponse"
								}
							}
						}
					}
				},
				"description": "This function will be called by the client application to receive an access and refresh token\nusing the authorization code from the login procedure.",
				"security": [],
				"parameters": [
					{
						"description": "The authorization code.",
						"in": "query",
						"name": "code",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "The redirect_uri for which the authorization code is valid. Must match the one supplied to {@link login}.",
						"in": "query",
						"name": "redirect_uri",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/spotify/refresh": {
			"get": {
				"operationId": "Refresh",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AccessTokenResponse"
								}
							}
						}
					}
				},
				"description": "Request a new access token using a previously obtained refresh token.",
				"security": [],
				"parameters": [
					{
						"description": "The refresh token obtained by calling {@link token}.",
						"in": "query",
						"name": "refresh_token",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}